separando el model.py
Me esta empezando a funcionar
