+++ model/db_config.py
@@ -1,3 +1,21 @@
+from genericpath import exists
+from peewee import *
+from model.class_product import Product
+
+database = SqliteDatabase('prueba.db')
+
+database.connect()
+lista_tablas = []
+
+if not(database.table_exists(Product)):
+    lista_tablas.append(Product)
+
+if (lista_tablas != []):
+    database.create_tables(lista_tablas)
+
+
+"""
+Original
 import sqlite3
 
 
@@ -8,7 +26,7 @@ def conexion_db():
 
 # Tabla
 def create_tablas(conexion):
-    create_query = """
+    create_query = """"""
                     CREATE TABLE IF NOT EXISTS producto(
                         id INTEGER PRIMARY KEY AUTOINCREMENT,
                         tipo varchar(50) NOT NULL,
@@ -16,7 +34,7 @@ def create_tablas(conexion):
                         referencia varchar(255) NOT NULL,
                         fecha_insert DATETIME NOT NULL
                     )
-                    """
+                    """"""
     cursor = conexion.cursor()
     cursor.execute(create_query)
     conexion.commit()
@@ -24,9 +42,9 @@ def create_tablas(conexion):
 
 
 def get_registros():
-    query = """
+    query = """"""
             SELECT * FROM producto
-            """
+            """"""
     conexion = conexion_db()
     cursor = conexion.cursor()
     cursor.execute(query)
@@ -40,10 +58,10 @@ def get_registros():
 
 def get_registro_by_id(id_producto):
     data_id = str(id_producto)
-    query = """
+    query = """"""
             SELECT * FROM producto
             WHERE id = ?
-            """
+            """"""
     conexion = conexion_db()
     cursor = conexion.cursor()
     cursor.execute(query, (data_id,))
@@ -58,11 +76,11 @@ def get_registro_by_id(id_producto):
 # Inserts
 def insert_producto(tipo, modelo, referencia, fecha):
     data = (tipo, modelo, referencia, fecha)
-    query = """
+    query = """"""
             INSERT INTO producto
             (tipo,modelo,referencia,fecha_insert)
             VALUES (?,?,?,?)
-            """
+            """"""
     conexion = conexion_db()
     cursor = conexion.cursor()
     cursor.execute(query, data)
@@ -78,13 +96,13 @@ def insert_producto(tipo, modelo, referencia, fecha):
 # Updates
 def update_producto(tipo, modelo, referencia, id):
     data = (tipo, modelo, referencia, id)
-    query = """
+    query = """"""
             UPDATE producto
             SET     tipo = ?,
                     modelo = ?,
                     referencia = ?
             WHERE id = ?
-            """
+            """"""
 
     conexion = conexion_db()
     cursor = conexion.cursor()
@@ -97,10 +115,10 @@ def update_producto(tipo, modelo, referencia, id):
 # Delete
 def delete_producto(id_producto):
     data = (id_producto,)
-    query = """
+    query = """"""
             DELETE FROM producto
             WHERE id = ?;
-            """
+            """"""
 
     conexion = conexion_db()
     cursor = conexion.cursor()
@@ -108,3 +126,4 @@ def delete_producto(id_producto):
 
     conexion.commit()
     conexion.close()
+"""
